Adding start to module selection, glow and menu
Can select a module and the game will recorgnise what it has selected.
Started adding in a glow around the module upon selection however it is permanantly on and will display throughother objects.
Also started the menu, there is a current problem where upon play the game will spawn in 2 menu gameobjects (this is mostlikely to do with the original module spawinging one to start with then another one is spawned upon game start). In the code need to add in a way to turn the menu on and off durning selection and not when the mouse is over the object.