BIG UPDATE
Modules have been seperated into empty gameobjects,
Implimented a way for the game to recorgnise what it is pointing at.
Got a text file to load into the game and stored the names individually.
Got the stored names to randomly create 10 staff members.
Cleaned up the componants so adding scripts is easier to find, mostly everything is now under "VRSS".